Add routing class
The routing class can implement ECMP or k-shortest paths for the topology.
Both at present use 8 for the max number of paths.
Routing creates a map from node id -> node id -> egress ports. Meaning, given the id
of the current node and the id of the destination node, we have a list of egress
ports for the paths. This should not remove duplicates (if there are multiple shortest
paths out of the same egress port, that egress port should appear in the list multiple times)