Figure 9
Adds code to generate figure 9. Takes about 1.5 mins to run.
Cleans up jellyfish topology generator so that it doesn't take
a ridiculous amount of time to run. Adds code to the routing class
to generate paths for random permutation traffic and count the number
of times each link is crossed in the paths. Found Yen's networkx
implementation and modifies it so that it actually works.

Running `python figure9.py` generates the graph and saves it to `figure9.png`