Use routing to set egress ports
Maintains list of paths between each pair of hosts and uses it to
set the egress port by finding a path from the src/dst mac of the packet,
finding where the current router is on the path, and determining out of which
port the next hop on the path is.

Currently does not use proper ecmp hashing and randomly chooses a path
because the riplpox ecmp hash doesn't work out of the box. We need to do this.

The dummy topology can pingall, but the Jellyfish topology is struggling
apparently because of DHCP packets being too short? I have been trying to
manually set IPs of switches to no avail.